+++ SensateIoT.SmartEnergy.Dsmr.Api/Data/LogoutRequest.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace SensateIoT.SmartEnergy.Dsmr.Api.Data
+{
+	public class LogoutRequest
+	{
+		[JsonProperty("email")]
+		public string Email { get; set; }
+	}
+}