+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Logout HTTP route
+- OTP request route
+
+### Updated
+- The authentication flow
+
 ## [1.3.0] - 27-04-2021
 ### Added
 - Hourly power data aggregate endpoint